% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetDelta}
\alias{GetDelta}
\title{Get quantity Delta.}
\usage{
GetDelta(nu, E.op, g.op, s.op)
}
\arguments{
\item{nu}{[\code{numeric}]\cr
A vector of the quantity nu.}

\item{E.op}{[\code{numeric}]\cr
A vector of expected values E(mu, mu.op, phi.op) on the Glicko-2 scale.}

\item{g.op}{[\code{numeric}]\cr
A vector of g(phi.op) on the Glicko-2 scale.}

\item{s.op}{[\code{numeric}]\cr
A vector of results \code{s.op} coded as (1 = win, 0 = loss, 0.5 = draw).}
}
\value{
[\code{numeric}] the numeric value of the quantity Delta.
}
\description{
Get quantity Delta.
}
