% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetE}
\alias{GetE}
\title{Get Expected value E(mu, mu.op, phi.op) for computing nu.}
\usage{
GetE(mu, mu.op, phi.op)
}
\arguments{
\item{mu}{[\code{numeric(1)}]\cr
A value of a players rating on the Glicko-2 scale.}

\item{mu.op}{[\code{numeric}]\cr
A vector of opponents ratings on the Glicko-2 scale.}

\item{phi.op}{[\code{numeric}]\cr
A vector of opponents rating deviations on the Glicko-2 scale.}
}
\value{
[\code{numeric}] the vector of E(mu, mu_j, phi_j).
}
\description{
Get Expected value E(mu, mu.op, phi.op) for computing nu.
}
