% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetG}
\alias{GetG}
\title{Get g(phi) for computing nu.}
\usage{
GetG(phi)
}
\arguments{
\item{phi}{[\code{numeric}]\cr
A vector of rating deviations on the Glicko-2 scale.}
}
\value{
[\code{numeric}] the vector of g(phi).
}
\description{
Get g(phi) for computing nu.
}
