% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetGlicko2Rating}
\alias{GetGlicko2Rating}
\title{Get a Glicko-2 rating update for a player.}
\usage{
GetGlicko2Rating(r = 1500, RD = 350, sigma = 0.06, r.op, RD.op, s.op,
  tau = 0.5, epsilon = 1e-06)
}
\arguments{
\item{r}{[\code{numeric(1)}]\cr
A players current rating r.}

\item{RD}{[\code{numeric(1)}]\cr
A players current rating deviation (RD).}

\item{sigma}{[\code{numeric(1)}]\cr
A players rating volatility sigma.}

\item{r.op}{[\code{numeric}]\cr
A vector of opponents current ratings r.}

\item{RD.op}{[\code{numeric}]\cr
A vector of opponents current rating deviations RD.}

\item{s.op}{[\code{numeric}]\cr
A vector of results s.op coded as (1 = win, 0 = loss, 0.5 = draw)
from players perspective.}

\item{tau}{[\code{numeric(1)}]\cr
A numeric value of the system constant tau.}

\item{epsilon}{[\code{numeric(1)}]\cr
A parameter used to measure convergance of the algorithm.}
}
\value{
A named [\code{list(3)}] with updated parameters of player
  strength, \code{r.new}, \code{RD.new}, \code{sigma.new}.
}
\description{
Get a Glicko-2 rating update for a player.
}
\examples{
# initial rating
e.r <- 1500
# initial rating deviation
e.RD <- 200
# initial volatility
e.sigma <- 0.06

# opponents ratings
e.r.op <- c(1400, 1550, 1700)
# opponents rating deviations
e.RD.op <- c(30, 100, 300)
# results
e.s.op <- c(1, 0, 0)  # win, loss, loss
# system constant
e.tau <- 0.5

GetGlicko2Rating(e.r, e.RD, e.sigma, e.r.op, e.RD.op, e.s.op, 0.5, 1e-6)
}
