% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetMu}
\alias{GetMu}
\title{Convert rating r to mu on Glicko-2 scale.}
\usage{
GetMu(r)
}
\arguments{
\item{r}{[\code{numeric}]\cr
A vector of ratings.}
}
\value{
[\code{numeric}] of ratings on the Glicko-2 scale.
}
\description{
Convert rating r to mu on Glicko-2 scale.
}
