% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetMuNew}
\alias{GetMuNew}
\title{Get the updated value of mu: mu.new.}
\usage{
GetMuNew(mu, phi.new, nu, Delta)
}
\arguments{
\item{mu}{[\code{numeric}]\cr
A vector of ratings on the Glicko-2 scale.}

\item{phi.new}{[\code{numeric}]\cr
A vector of updated rating volatilities on the Glicko-2 scale.}

\item{nu}{[\code{numeric}]\cr
A vector of the quantity nu.}

\item{Delta}{[\code{numeric}]\cr
A vector of the quantity Delta.}
}
\value{
[\code{numeric}] A vector of the updated values of mu: mu.new.
}
\description{
Get the updated value of mu: mu.new.
}
