% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetNu}
\alias{GetNu}
\title{Get quantity nu.}
\usage{
GetNu(E.op, g.op)
}
\arguments{
\item{E.op}{[\code{numeric}]\cr
A vetor of expected values E(mu, mu.op, phi.op) on the Glicko-2 scale.}

\item{g.op}{[\code{numeric}]\cr
A vector of g(phi.op) on the Glicko-2 scale.}
}
\value{
[\code{numeric(1)}] the numeric value of the quantity nu.
}
\description{
Get quantity nu.
}
