% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetPhi}
\alias{GetPhi}
\title{Convert rating deviation RD to phi on Glicko-2 scale.}
\usage{
GetPhi(RD)
}
\arguments{
\item{RD}{[\code{numeric}]\cr
A vector of rating deviations (RD).}
}
\value{
[\code{numeric}] the numeric vector of rating deviations on
  the Glicko-2 scale.
}
\description{
Convert rating deviation RD to phi on Glicko-2 scale.
}
