% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetPhiNew}
\alias{GetPhiNew}
\title{Get the updated value of phi: phi.new.}
\usage{
GetPhiNew(phi.pre, sigma.new, nu)
}
\arguments{
\item{phi.pre}{[\code{numeric}]\cr
A vector of pre-rating deviations on the Glicko-2 scale.}

\item{sigma.new}{[\code{numeric}]\cr
A vector of the updated rating volatility on the Glicko-2 scale.}

\item{nu}{[\code{numeric(1)}]\cr
A numeric value of the quantity nu.}
}
\value{
[\code{numeric(1)}] a vector of the updated values of phi: phi.new.
}
\description{
Get the updated value of phi: phi.new.
}
