% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetPhiPre}
\alias{GetPhiPre}
\title{Get the updated value of phi to new pre-rating period: phi.pre.}
\usage{
GetPhiPre(phi, sigma.new)
}
\arguments{
\item{phi}{[\code{numeric}]\cr
A vector of the rating deviations on the Glicko-2 scale.}

\item{sigma.new}{[\code{numeric}]\cr
A vector of the updated rating volatility for a given player on the
Glicko-2 scale.}
}
\value{
[\code{numeric}] a vector of the updated value of phi to
  new pre-rating period.
}
\description{
Get the updated value of phi to new pre-rating period: phi.pre.
}
