% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetRDNew}
\alias{GetRDNew}
\title{Get the updated value of RD: RD.new.}
\usage{
GetRDNew(phi.new)
}
\arguments{
\item{phi.new}{[\code{numeric}]\cr
A vector of the rating volatility on the Glicko-2 scale.}
}
\value{
[\code{numeric}] a vector of the updated value of the rating RD.
}
\description{
Get the updated value of RD: RD.new.
}
