% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetRNew}
\alias{GetRNew}
\title{Get the updated value of r: r.new.}
\usage{
GetRNew(mu.new)
}
\arguments{
\item{mu.new}{[\code{numeric}]\cr
A vector of updated ratings on the Glicko-2 scale.}
}
\value{
[\code{numeric}] A vector of the updated values of the rating r.
}
\description{
Get the updated value of r: r.new.
}
