% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glicko2Functions.R
\name{GetSigmaNew}
\alias{GetSigmaNew}
\title{Get the updated value of sigma: sigma.new.}
\usage{
GetSigmaNew(sigma, phi, Delta, nu, tau, epsilon)
}
\arguments{
\item{sigma}{[\code{numeric(1)}]\cr
A numeric value of the rating volatility for a given player on the
Glicko-2 scale.}

\item{phi}{[\code{numeric(1)}]\cr
A numeric value of the rating deviation on the Glicko-2 scale.}

\item{Delta}{[\code{numeric(1)}]\cr
A numeric value of quantity Delta.}

\item{nu}{[\code{numeric(1)}]\cr
A numeric value of the quantity nu.}

\item{tau}{[\code{numeric(1)}]\cr
A numeric value of the system constant tau.}

\item{epsilon}{[\code{numeric(1)}]\cr
A parameter used to measure convergance of the algorithm.}
}
\value{
[\code{numeric(1)}] the updated value of \code{sigma}.
}
\description{
Get the updated value of sigma: sigma.new.
}
