% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeData.R
\name{InitializeData}
\alias{InitializeData}
\title{Initialize data (matches and strengths) for Glicko-2}
\usage{
InitializeData(player, opponent, is_winner, period, is_symmetric = FALSE,
  initial_r = 1500, initial_RD = 350, initial_sigma = 0.06)
}
\arguments{
\item{player}{[\code{charcater}] or [\code{integer}]\cr
Players, either by character or by index.}

\item{opponent}{[\code{charcater}] or [\code{integer}]\cr
Opponents of players, either by character or by index.}

\item{is_winner}{[\code{numeric}] or [\code{logical}]\cr
Indicator wheter \code{player} won against \code{opponent}.
If \code{numeric} must only containt ones and zeros.}

\item{period}{[\code{integer}]\cr
Period the match of \code{player} vs. \code{opponent}.}

\item{is_symmetric}{[\code{logical(1)}]\cr
Whether data is already provided symmetrically. E.g. if one match was
played and \code{is_symmetric=TRUE} there must exist two entries in
all supplied vectors, one for the first player as player and the second
player as opponent and vice versa (see examples below).}

\item{initial_r}{[\code{numeric(0)}]\cr
Initial rating for Glicko-2.}

\item{initial_RD}{[\code{numeric(0)}]\cr
Initial rating deviance for Glicko-2.}

\item{initial_sigma}{[\code{numeric(0)}]\cr
Initial sigma for Glicko-2.}
}
\value{
A named [\code{list(2)}] with a [\code{data.table}] object
  containing players strengths (\code{strengths}) with columns
  \code{player, period, r, RD, sigma} and a [\code{data.table}] object
  containing a symmetric set of matches with columns (\code{mactches})
  \code{player, opponent, is_winner, period}
}
\description{
Initialize data (matches and strengths) for Glicko-2
}
\examples{
InitializeData(player = c("A", "B", "C"),
               opponent = c("D", "D", "D"),
               is_winner = c(1, 0, 0),
               period = c(1, 1, 2))
}
