% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeData.R
\name{InitializeMatches}
\alias{InitializeMatches}
\title{Initialize  matches for Glicko-2}
\usage{
InitializeMatches(player, opponent, is_winner, period, is_symmetric = FALSE)
}
\arguments{
\item{player}{[\code{charcater}] or [\code{integer}]\cr
Players, either by character or by index.}

\item{opponent}{[\code{charcater}] or [\code{integer}]\cr
Opponents of players, either by character or by index.}

\item{is_winner}{[\code{numeric}] or [\code{logical}]\cr
Indicator wheter \code{player} won against \code{opponent}.
If \code{numeric} must only containt ones and zeros.}

\item{period}{[\code{integer}]\cr
Period the match of \code{player} vs. \code{opponent}.}

\item{is_symmetric}{[\code{logical(1)}]\cr
Whether data is already provided symmetrically. E.g. if one match was
played and \code{is_symmetric=TRUE} there must exist two entries in
all supplied vectors, one for the first player as player and the second
player as opponent and vice versa (see examples below).}
}
\value{
A [\code{data.table}] object containing a symmetric
set of matches with columns \code{player, opponent, is_winner, period}.
}
\description{
Initialize  matches for Glicko-2
}
\examples{
InitializeMatches(player = c("A", "B", "C"),
                  opponent = c("D", "D", "D"),
                  is_winner = c(1, 0, 0),
                  period = c(1, 1, 2))
}
