% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeData.R
\name{InitializeStrengths}
\alias{InitializeStrengths}
\title{Initialize  strengths for Glicko-2}
\usage{
InitializeStrengths(player, period = NULL, r = NULL, RD = NULL,
  sigma = NULL, initial_r = 1500, initial_RD = 350,
  initial_sigma = 0.06)
}
\arguments{
\item{player}{[\code{charcater}] or [\code{integer}]\cr
of players, either by character or by index. Must be of
\code{length(period)} is \code{period} is given.}

\item{period}{[\code{integer}]\cr
period indices.}

\item{r}{[\code{numeric}]\cr
ratings from Glicko-2.}

\item{RD}{\code{numeric}]\cr
rating deviations RD from Glicko-2.}

\item{sigma}{[\code{numeric}]\cr
sigma from Glicko-2.}

\item{initial_r}{[\code{numeric(0)}]\cr
Initial rating for Glicko-2.}

\item{initial_RD}{[\code{numeric(0)}]\cr
Initial rating deviance for Glicko-2.}

\item{initial_sigma}{[\code{numeric(0)}]\cr
Initial sigma for Glicko-2.}
}
\value{
A [\code{data.table}] object containing players
 strengths with columns \code{player, period, r, RD, sigma}.
}
\description{
Initialize  strengths for Glicko-2
}
\examples{
InitializeStrengths(player = c("Foo", "Bar", "Baz"))
}
