% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdatePlayerRating.R
\name{UpdatePlayerRating}
\alias{UpdatePlayerRating}
\title{Update a single player's rating for a given period.}
\usage{
UpdatePlayerRating(selected_player, selected_period, period_strengths,
  period_matches, tau = 0.5, epsilon = 1e-06)
}
\arguments{
\item{selected_player}{[\code{chracter(1)}]\cr
Name of the player for whom updates should be calculated.}

\item{selected_period}{[\code{numeric(1)}]\cr
Period for which updates should be calculated.}

\item{period_strengths}{[\code{data.table}]\cr
A \code{data.table} object containing players strengths with columns
\code{player, period, r, RD, sigma} for selected player, opponents
and selected period (\code{selected_period}).}

\item{period_matches}{[\code{data.table}]\cr
A \code{data.table} object containing matches with columns
\code{player, period, opponent, is_winner}.}

\item{tau}{[\code{numeric(1)}]\cr
Parameter \code{tau} for the Glicko-2 algorithm, reasonable choices are
in the interval \code{[0.3, 1.2]}.}

\item{epsilon}{[\code{numeric(1)}]\cr
Parameter \code{epsilon} for the Glicko-2 algorithm.}
}
\value{
A [\code{data.table}] object containing updates for selected players
  \code{selected_player} strength with columns
  \code{player, period, r, RD, sigma}.
}
\description{
Update a single player's rating for a given period.
}
