% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdatePlayerRatingPeriods.R
\name{UpdatePlayerRatingPeriods}
\alias{UpdatePlayerRatingPeriods}
\title{Update ratings for a given number of periods and players.}
\usage{
UpdatePlayerRatingPeriods(strengths, matches,
  max_period = max(matches$period), min_period = max(strengths$period) + 1,
  tau = 0.5, epsilon = 1e-06)
}
\arguments{
\item{strengths}{[\code{data.table}]\cr
A \code{data.table} object containing players strengths with columns
\code{player, period, r, RD, sigma}.}

\item{matches}{[\code{data.table}]\cr
A \code{data.table} object containing matches with columns
\code{player, period, opponent, is_winner}.}

\item{max_period}{[\code{numierc(1)}]\cr Maximum period up to which
and including rating updates are calculated.}

\item{min_period}{[\code{numierc(1)}]\cr Minimum period from which
and including rating updates are calculated.}

\item{tau}{[\code{numeric(1)}]\cr
Parameter \code{tau} for the Glicko-2 algorithm, reasonable choices are in \code{[0.3, 1.2]}.}

\item{epsilon}{[\code{numeric(1)}]\cr
Parameter \code{epsilon} for the Glicko-2 algorithm.}
}
\value{
A [\code{data.table}] object containing updates on
 strength with columns \code{player, period, r, RD, sigma} for all
 players and the given (\code{strengths}) and selected range of periods
 (\code{min_period}, \code{max_period}).
}
\description{
Update ratings for a given number of periods and players.
}
\examples{
data <- InitializeData(player = c("A", "B", "C"),
                       opponent = c("D", "D", "D"),
                       is_winner = c(1, 0, 0),
                       period = c(1, 1, 2))

UpdatePlayerRatingPeriods(data$strengths, data$matches)
}
